sigma_warehouse/requirements.txt
Django==5.0.2
djangorestframework==3.14.0
drf-yasg==1.21.7
psycopg2-binary==2.9.9
python-dotenv==1.0.1